# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "eShop - Catalog HTTP API",
    "description": "The Catalog Microservice HTTP API. This is a Data-Driven/CRUD microservice sample",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:62499/"
    }
  ],
  "paths": {
    "/api/catalog/items/by": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Batch get catalog items",
        "description": "Get multiple items from the catalog",
        "operationId": "BatchGetItems",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "List of ids for catalog items to return",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "pattern": "^-?(?:0|[1-9]\\d*)$",
                "type": [
                  "integer",
                  "string"
                ],
                "format": "int32"
              }
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "description": "The API version, in the format 'major.minor'.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1.0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CatalogItem"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalog/items/{id}": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Get catalog item",
        "description": "Get an item from the catalog",
        "operationId": "GetItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The catalog item id",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "description": "The API version, in the format 'major.minor'.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1.0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItem"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Catalog"
        ],
        "summary": "Delete catalog item",
        "description": "Delete the specified catalog item",
        "operationId": "DeleteItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the catalog item to delete",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "description": "The API version, in the format 'major.minor'.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1.0"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/catalog/items/{id}/pic": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Get catalog item picture",
        "description": "Get the picture for a catalog item",
        "operationId": "GetItemPicture",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The catalog item id",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "description": "The API version, in the format 'major.minor'.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1.0"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              },
              "image/gif": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              },
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              },
              "image/bmp": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              },
              "image/tiff": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              },
              "image/wmf": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              },
              "image/jp2": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              },
              "image/svg+xml": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              },
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalog/catalogtypes": {
      "get": {
        "tags": [
          "Types"
        ],
        "summary": "List catalog item types",
        "description": "Get a list of the types of catalog items",
        "operationId": "ListItemTypes",
        "parameters": [
          {
            "name": "api-version",
            "in": "query",
            "description": "The API version, in the format 'major.minor'.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1.0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CatalogType"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalog/catalogbrands": {
      "get": {
        "tags": [
          "Brands"
        ],
        "summary": "List catalog item brands",
        "description": "Get a list of the brands of catalog items",
        "operationId": "ListItemBrands",
        "parameters": [
          {
            "name": "api-version",
            "in": "query",
            "description": "The API version, in the format 'major.minor'.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1.0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CatalogBrand"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/catalog/items/by/{name}": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Get catalog items by name",
        "description": "Get a paginated list of catalog items with the specified name.",
        "operationId": "GetItemsByName",
        "parameters": [
          {
            "name": "PageSize",
            "in": "query",
            "description": "Number of items to return in a single page of results",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "PageIndex",
            "in": "query",
            "description": "The index of the page of results to return",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "The name of the item to return",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "description": "The API version, in the format 'major.minor'.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1.0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedItemsOfCatalogItem"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CatalogBrand": {
        "required": [
          "brand"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "brand": {
            "type": "string"
          }
        }
      },
      "CatalogItem": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "price": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "pictureFileName": {
            "type": [
              "null",
              "string"
            ]
          },
          "catalogTypeId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "catalogType": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CatalogType"
              }
            ]
          },
          "catalogBrandId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "catalogBrand": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CatalogBrand"
              }
            ]
          },
          "availableStock": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "restockThreshold": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "maxStockThreshold": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "onReorder": {
            "type": "boolean"
          }
        }
      },
      "CatalogType": {
        "required": [
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      }
    }
  },
...
}